requirements.txt
telethon
python-dotenv
pydantic
pydantic-settings
clamd
